검색창 디자인 정리 및 뷰모델 추가
