+++ .idea/misc.xml
@@ -3,6 +3,7 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="..\:/JOB/Android/Jejudo/app/src/main/res/drawable/square_border.xml" value="0.1585" />
         <entry key="..\:/JOB/Android/Jejudo/app/src/main/res/layout/activity_main.xml" value="0.16666666666666666" />
         <entry key="..\:/JOB/Android/Jejudo/app/src/main/res/layout/cell_festival.xml" value="0.20364583333333333" />
         <entry key="..\:/JOB/Android/Jejudo/app/src/main/res/layout/cell_festival_forsearch.xml" value="0.33" />
