+++ app/src/main/java/com/example/teamjejudo/MainActivity.kt
@@ -21,7 +21,6 @@ class MainActivity : AppCompatActivity() {
         )
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
-
     }
 
     override fun onSupportNavigateUp(): Boolean {
