+++ app/src/main/java/com/example/teamjejudo/screen/festival/FestivalDetailFragment.kt
@@ -24,10 +24,8 @@ class FestivalDetailFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-
         _binding = FragmentFestivalDetailBinding.inflate(inflater, container, false)
         return binding.root
-
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
