+++ app/src/main/res/drawable/square_border.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+    <item>
+        <shape android:shape="rectangle" >
+            <solid android:color="#9E9898"/>
+            <corners android:radius="16dp"/>
+        </shape>
+    </item>
+    <item android:bottom="0.5dp" android:top="0.5dp" android:start="0.5dp" android:end="0.5dp">
+        <shape android:shape="rectangle" >
+            <solid android:color="#FFFFFF" />
+            <corners android:radius="16dp"/>
+        </shape>
+    </item>
+</layer-list>
