+++ app/src/main/res/layout/fragment_festival.xml
@@ -16,21 +16,27 @@
 
     <EditText
         android:id="@+id/et_searchbar"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintLeft_toLeftOf="parent"
         android:layout_width="match_parent"
         android:layout_height="35dp"
-        android:layout_marginTop="10dp"
-        android:layout_marginLeft="15dp"
-        android:layout_marginRight="15dp"
+        android:layout_marginLeft="12dp"
+        android:layout_marginRight="12dp"
+        android:layout_marginTop="8dp"
+        android:background="@drawable/square_border"
+        android:hint="축제 검색"
         android:padding="5dp"
-        android:textSize="20dp"
-        android:hint="축제 검색" />
+        android:textSize="15sp"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rv_festival_forsearch"
         android:layout_width="match_parent"
         android:layout_height="0dp"
+        android:background="@color/white"
+        android:elevation="2dp"
+        android:layout_marginLeft="15dp"
+        android:layout_marginRight="15dp"
+        android:layout_toStartOf="@+id/tv_ongoing_festival_title"
         app:layout_constraintTop_toBottomOf="@id/et_searchbar"
         app:layout_constraintLeft_toLeftOf="parent"
         app:isGone="@{!isFocused}"/>
@@ -60,9 +66,7 @@
         android:paddingBottom="10dp"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/tv_ongoing_festival_title"
-
-        />
+        app:layout_constraintTop_toBottomOf="@id/tv_ongoing_festival_title"     />
 
     <Button
         android:id="@+id/button_first"
